requirements.txt
openai
yfinance
pandas
openpyxl
python-dotenv
